extract GuiGlobalMessageEntryView
