+++ CMakeLists.txt
@@ -128,6 +128,7 @@ set(SOURCES
 	src/screens/extra/shipLogScreen.cpp
 	src/screens/gm/gameMasterScreen.cpp
 	src/screens/gm/objectCreationView.cpp
+	src/screens/gm/globalMessageEntryView.cpp
 	src/screens/gm/chatDialog.cpp
 	src/screens/gm/tweak.cpp
 	src/screenComponents/aimLock.cpp
