+++ src/screens/gm/gameMasterScreen.h
@@ -20,6 +20,7 @@ class GuiToggleButton;
 class GuiTextEntry;
 class GameMasterChatDialog;
 class GuiObjectCreationView;
+class GuiGlobalMessageEntryView;
 class GameMasterScreen : public GuiCanvas, public Updatable
 {
 private:
@@ -31,7 +32,7 @@ private:
     
     GuiElement* chat_layer;
     std::vector<GameMasterChatDialog*> chat_dialog_per_ship;
-    GuiGlobalMessageEntry* global_message_entry;
+    GuiGlobalMessageEntryView* global_message_entry;
     GuiObjectCreationView* object_creation_view;
     GuiObjectTweak* player_tweak_dialog;
     GuiObjectTweak* ship_tweak_dialog;
@@ -79,14 +80,5 @@ public:
     string getScriptExport(bool selected_only);
 };
 
-class GuiGlobalMessageEntry : public GuiOverlay
-{
-private:
-    GuiTextEntry* message_entry;
-public:
-    GuiGlobalMessageEntry(GuiContainer* owner);
-    
-    virtual bool onMouseDown(sf::Vector2f position);
-};
 
 #endif//GAME_MASTER_SCREEN_H
