+++ src/screens/gm/globalMessageEntryView.cpp
@@ -1,0 +1,35 @@
+#include "globalMessageEntryView.h"
+#include "gameGlobalInfo.h"
+
+#include "gui/gui2_panel.h"
+#include "gui/gui2_textentry.h"
+#include "gui/gui2_button.h"
+
+GuiGlobalMessageEntryView::GuiGlobalMessageEntryView(GuiContainer* owner)
+: GuiOverlay(owner, "GLOBAL_MESSAGE_ENTRY", sf::Color(0, 0, 0, 128))
+{
+    GuiPanel* box = new GuiPanel(this, "FRAME");
+    box->setPosition(0, 0, ACenter)->setSize(800, 150);
+    
+    message_entry = new GuiTextEntry(box, "MESSAGE_ENTRY", "");
+    message_entry->setPosition(0, 20, ATopCenter)->setSize(700, 50);
+    
+    (new GuiButton(box, "CLOSE_BUTTON", "Cancel", [this]() {
+        this->hide();
+    }))->setPosition(20, -20, ABottomLeft)->setSize(300, 50);
+
+    (new GuiButton(box, "SEND_BUTTON", "Send", [this]() {
+        string message = message_entry->getText();
+        if (message.length() > 0)
+        {
+            gameGlobalInfo->global_message = message;
+            gameGlobalInfo->global_message_timeout = 5.0;
+        }
+        this->hide();
+    }))->setPosition(-20, -20, ABottomRight)->setSize(300, 50);
+}
+
+bool GuiGlobalMessageEntryView::onMouseDown(sf::Vector2f position)
+{   //Catch clicks.
+    return true;
+}
